feat: simplified setup page and elevation ranges  (#75)
Simplify the setup page: The user selects the data she wants to import to DHIS2, and then we only show the data element, data set, data element group and legend that is needed.

Elevation distribution chart: Group elevation values (every 10 or 100 meters) if the elevation range is more than 200 meters. It will make the elevation distribution chart smoother and easier to read, and make sure the chart is showing when the elevation range is thousands of meters (e.g. Nepal).