feat: import page improvements (#89)
https://dhis2.atlassian.net/browse/CLIM-17

In this PR, the import page has been restructured and the UX improved:

* The data element list is populated based on the selected period type.
This happens by fetching all the data sets and filtering based on the selected period type,
resulting in a list of data elements for the period type. The app ignores if a data element
has been added to 2 or more data sets having different period types. It is up to the implementation
properly structure the metadata. The setup guide should help admins avoid this situation,
but we can't guarantee this.

* The instructions on the side have been removed, and instead, the import form
includes the needed info directly in the form.

* Visual styling improvements

* The period type selector is placed above other selections because it influences
what data elements are available in the data element selector.

* If it is relevant to either know or select the time zone (e.g. hourly data), then it is displayed.

* The setup guide now recommends including the selected period type in the code
of the data element (for autofilling the data element to import into dhis2).

* An import preview summary has been added to inform the user of what is being imported.

* Unit tests have been added to demonstrate support for non gregorian calendars
* Date validation based on date range for a dataset. It will inform the user if a date is
selected that is outside of the range.